straight moves of the various synthentic tests
